+++ js-miniapp-bridge/src/common-bridge.ts
@@ -36,22 +36,24 @@ import {
 } from './types/analytics/analytics';
 import { UniversalBridgeInfo } from './types/universal-bridge';
 import { CookieInfo } from './types/cookie-info';
-import { NotificationBridge } from './modules/notification-bridge';
 import {
   NotificationDetailedInfo,
   NotificationInfo,
   NotificationInfoType,
 } from './types/notification/notification-info';
+import { LogType } from './types/log-type';
+import { EsimConfig } from './types/e-sim';
+import { Platform } from './types/platform';
+import { LaunchBrowserOptions } from './types/browser-options';
+
+import { NotificationBridge } from './modules/notification-bridge';
 import { MiniAppPreferences } from './modules/miniapp-preferences';
 import { BrowserManager } from './modules/browser-manager';
 import { GalleryManager } from './modules/gallery-manager';
 import { UserProfileManager } from './modules/userprofile-manager';
 import { WebViewConfigManager } from './modules/webview-config-manager';
 import { UtilityManager } from './modules/utility-manager';
-import { LogType } from './types/log-type';
-import { EsimConfig } from './types/e-sim';
-import { Platform } from './types/platform';
-import { LaunchBrowserOptions } from './types/browser-options';
+import { OneClickSdk } from './modules/one-click-sdk';
 
 /** @internal */
 const mabMessageQueue: Callback[] = [];
@@ -108,6 +110,7 @@ export class MiniAppBridge {
   userProfileManager: UserProfileManager;
   webviewConfigManager: WebViewConfigManager;
   utilityManager: UtilityManager;
+  oneClickSdk: OneClickSdk;
 
   constructor(executor: PlatformExecutor) {
     this.executor = executor;
@@ -119,6 +122,7 @@ export class MiniAppBridge {
     this.userProfileManager = new UserProfileManager(executor);
     this.webviewConfigManager = new WebViewConfigManager(executor);
     this.utilityManager = new UtilityManager(executor);
+    this.oneClickSdk = new OneClickSdk(executor);
 
     if (window) {
       window.addEventListener(