+++ js-miniapp-bridge/src/index.ts
@@ -75,6 +75,7 @@ import { LogType } from './types/log-type';
 
 import { EsimConfig } from './types/e-sim';
 
+export { OneClickSdkInfo } from './types/one-click-sdk';
 export { MAAnalyticsConfig } from './types/analytics/analytics';
 
 export {