+++ js-miniapp-bridge/src/types/one-click-sdk.ts
@@ -1,0 +1,7 @@
+export interface OneClickSdkInfo {
+  idid: string;
+  minor: boolean;
+  redirectUri: string;
+  supportedKycTypes: string;
+  enabledSecurityCheck?: boolean;
+}